rename(constants): rename magicLiterals to constants and add comments
