+++ src/utils/constants.js
@@ -1,9 +1,7 @@
+/* constants for number of items */
 export const MOBILE = 0;
 export const TABLET = 1;
 export const PC = 2;
-
-export const BASE_API_URL = "https://panda-market-api.vercel.app/";
-export const PRODUCT_API_URL = `${BASE_API_URL}products/`;
 
 export const NUM_BEST_ITEMS = {
   [MOBILE]: 1,
@@ -17,6 +15,7 @@ export const NUM_ALL_ITEMS = {
   [PC]: 10,
 };
 
+/* constants for order of items */
 export const RECENT = "recent";
 export const FAVORITE = "favorite";
 
@@ -24,3 +23,7 @@ export const ORDER_MESSAGE = {
   [FAVORITE]: "좋아요순",
   [RECENT]: "최신순",
 };
+
+/* constants for api url */
+export const BASE_API_URL = "https://panda-market-api.vercel.app/";
+export const PRODUCT_API_URL = `${BASE_API_URL}products/`;
